multidvorak: hwdvorak can be toggled off using the same key sequence
Previously the device needed to be removed from power to disable
hardware dvorak mode.  Now the mode-switching key (tab) toggles the
enabled mode.

Futhermore, switching layer orbits puts the user onto the new base layer
without needing to press fn.  This appears to be the intention of the
original code.  The current_layer was being set but the effect was not
seen because current_layer_number wasn't being set as well.